refactor(ComponentAttr): 重构属性设置面板，优化表单逻辑和组件渲染 refactor(ActionCard): 调整样式，增加右边距 refactor(EventItem): 修复事件项的样式问题 feat(FormElementRenderer): 新增表单元素渲染器，支持动态控件渲染 feat(useComponentAttrForm): 添加自定义 Hook 管理属性面板表单逻辑 feat(formUtils): 增强表单工具函数，支持选项归一化和动态属性名获取
